vaults-backend/pyproject.toml
[tool.poetry]
name = "src"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
#readme = "README.md"


[tool.poetry.dependencies]
python = "^3.10" 
anchorpy = "0.20.1"
solana = "^0.34.0"
requests = "^2.28.1"
#types-requests = "^2.28.9"
ccxt = "^4.0.0"
pandas = "^1.5.3"
pandas-ta = "^0.3.14b0"
anchorpy-core = "0.2.0"
python-dotenv = "^1.0.0"
#drift-jit-proxy = "0.1.4"
#asyncio = "^3.4.3"
driftpy = "0.8.16"
Backtesting = "^0.3.3"
matplotlib = "^3.9.2"
seaborn = "^0.13.2"

web3 = "^6.0.0"
#aiohttp = "^3.9.1"
#websockets = "^12.0"
eth-abi = "^4.2.1"
eth-utils = "^2.3.1"
eth-account = "^0.9.0"
degenbot = "^0.2.0"
ujson = "^5.9.0"


[tool.poetry.dev-dependencies]
pytest = "^7.2.0"
pytest-asyncio = "^0.21.0"
pytest-xprocess = "0.18.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
